Move middleware into individual middleware/* files
